Added 404s for non-existant id's

+++ tvm/src/tvm/text_endpoints.py
@@ -11,19 +11,21 @@ from datetime import datetime
 @app.get("/categories/")
 def read_categories(db: Session = Depends(get_db)):
     categories = db.query(Category).all()
+    if not categories:
+        return HTTPException(status_code=404, detail="No categories found.")
     return categories
 
 # Get the category with the given ID
 @app.get("/categories/{category_id}")
 def read_category(category_id: int, db: Session = Depends(get_db)):
-    # category = db.query(Category).filter(Category.id == category_id).first()
     category = db.get(Category, category_id)
+    if not category:
+        return HTTPException(status_code=404, detail="Category not found.")
     return category
 
 # Create a new category
 @app.post("/categories/")
 def create_category(category_name: str, db: Session = Depends(get_db)):
-    # category = db.query(Category).filter(Category.name == category_name).first()
     category = db.get(Category, category_name)
     if category:
         raise HTTPException(status_code=400, detail="This category already exists.")
@@ -36,8 +38,9 @@ def create_category(category_name: str, db: Session = Depends(get_db)):
 # Update the category with the given ID
 @app.put("/categories/{category_id}")
 def update_category(category_id: int, category_name: str, db: Session = Depends(get_db)):
-    # category = db.query(Category).filter(Category.id == category_id).first()
     category = db.get(Category, category_id)
+    if not category:
+        return HTTPException(status_code=404, detail="Category not found.")
     setattr(category, Category.text, category_name)
     db.commit()
     db.refresh(category)
@@ -46,8 +49,9 @@ def update_category(category_id: int, category_name: str, db: Session = Depends(
 # Delete the category with the given ID
 @app.delete("/categories/{category_id}")
 def delete_category(category_id: int, db: Session = Depends(get_db)):
-    # category = db.query(Category).filter(Category.id == category_id).first()
     category = db.get(Category, category_id)
+    if not category:
+        return HTTPException(status_code=404, detail="Category not found.")
     db.delete(category)
     db.commit()
     db.refresh(category)
@@ -57,12 +61,16 @@ def delete_category(category_id: int, db: Session = Depends(get_db)):
 @app.get("/subcategories/")
 def read_subcategories(db: Session = Depends(get_db)):
     subcategories = db.query(SubCategory).all()
+    if not subcategories:
+        return HTTPException(status_code=404, detail="No subcategories found.")
     return subcategories
 
 # Get the subcategory with the given ID
 @app.get("/subcategories/{subcategory_id}")
 def read_subcategory(subcategory_id: int, db: Session = Depends(get_db)):
     subcategory = db.query(SubCategory).filter(SubCategory.id == subcategory_id).first()
+    if not subcategory:
+        return HTTPException(status_code=404, detail="Subcategory not found.")
     return subcategory
 
 # Create a new subcategory
@@ -80,8 +88,9 @@ def create_subcategory(subcategory_name: str, db: Session = Depends(get_db)):
 # Update the subcategory with the given ID
 @app.put("/subcategories/{subcategory_id}")
 def update_subcategory(subcategory_id: int, subcategory_name: str, db: Session = Depends(get_db)):
-    # subcategory = db.query(SubCategory).filter(SubCategory.id == subcategory_id).first()
     subcategory = db.get(SubCategory, subcategory_id)
+    if not subcategory:
+        return HTTPException(status_code=404, detail="Subcategory not found.")
     setattr(subcategory, SubCategory.text, subcategory_name)
     db.commit()
     db.refresh(subcategory)
@@ -90,8 +99,9 @@ def update_subcategory(subcategory_id: int, subcategory_name: str, db: Session =
 # Delete the subcategory with the given ID
 @app.delete("/subcategories/{subcategory_id}")
 def delete_category(subcategory_id: int, db: Session = Depends(get_db)):
-    # subcategory = db.query(SubCategory).filter(SubCategory.id == subcategory_id).first()
     subcategory = db.get(SubCategory, subcategory_id)
+    if not subcategory:
+        return HTTPException(status_code=404, detail="Subcategory not found.")
     db.delete(subcategory)
     db.commit()
     db.refresh(subcategory)
@@ -101,17 +111,19 @@ def delete_category(subcategory_id: int, db: Session = Depends(get_db)):
 @app.get("/advisorytexts/")
 def read_texts(db: Session = Depends(get_db)):
     advisorytexts = db.query(AdvisoryText).all()
+    if not advisorytexts:
+        return HTTPException(status_code=404, detail="No advice texts found.")
     return advisorytexts
 
 # Gets the advice text with the given composite ID's
 @app.get("/advisorytexts/catid={category_id}&subid={subcategory_id}")
 def read_text(catid: int, subid: int, db: Session = Depends(get_db)):
     advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
+    if not advisorytext:
+        return HTTPException(status_code=404, detail="Advice text not found.")
     return advisorytext
 
 # Create new advice text
-### For the DB, there will be two dimensions: category and subcategory.
-### Advices table will be a composite table of both dimensions
 @app.post("/advisorytexts/")
 def create_text(advice_text: str, db: Session = Depends(get_db)):
     advice = db.query(AdvisoryText).filter(AdvisoryText.text == advice_text).first()
@@ -127,6 +139,8 @@ def create_text(advice_text: str, db: Session = Depends(get_db)):
 @app.put("/advisorytexts/catid={category_id}&subid={subcategory_id}")
 def update_text(catid: int, subid: int, advice_text: str, db: Session = Depends(get_db)):
     advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
+    if not advisorytext:
+        return HTTPException(status_code=404, detail="Advice text not found.")
     setattr(advisorytext, AdvisoryText.text, advice_text)
     db.commit()
     db.refresh(advisorytext)
@@ -136,6 +150,8 @@ def update_text(catid: int, subid: int, advice_text: str, db: Session = Depends(
 @app.delete("/advisorytexts/catid={category_id}&subid={subcategory_id}")
 def delete_text(catid: int, subid: int, db: Session = Depends(get_db)):
     advisorytext = db.query(AdvisoryText).filter(AdvisoryText.category_id == catid && AdvisoryText.subcategory_id == subid).first()
+    if not advisorytext:
+        return HTTPException(status_code=404, detail="Advice text not found.")
     db.delete(advisorytext)
     db.commit()
     db.refresh(advisorytext)
